pyproject.toml
[project]
name = "yt-video-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "fastapi[standard]>=0.128.0",
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.3",
    "google-genai>=1.55.0",
    "httpx>=0.28.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "youtube-transcript-api>=1.2.3",
]
